style changes, components/directories reorganized
