+++ src/Components/Nav/Nav.css
@@ -43,8 +43,9 @@
 }
 
 .activeTabContent {
-  height: 70%;
+  height: fit-content;
   padding-top: 20px;
+  padding-bottom: 20px;
   border-top-right-radius: 5px;
   border-bottom-left-radius: 5px;
   border-bottom-right-radius: 5px;
@@ -55,6 +56,26 @@
   flex-direction: column;
   justify-content: flex-start;
   align-items: center;
+}
+
+.errorNotification {
+  width: 90%;
+  padding-top: 5px;
+  padding-bottom: 5px;
+  background-color: rgba(10, 187, 54, 0.637);
+  border-radius: 5px;
+  box-shadow: 1px 1px 5px rgb(102, 252, 3),
+    -1px -1px 5px rgba(253, 10, 253, 0.945);
+}
+
+.errorNotification {
+  width: 90%;
+  padding-top: 5px;
+  padding-bottom: 5px;
+  background-color: rgba(228, 13, 13, 0.445);
+  border-radius: 5px;
+  box-shadow: 1px 1px 5px rgba(0, 0, 0, 0.199),
+    -1px -1px 5px rgba(17, 17, 17, 0.199);
 }
 
 .tabLink {
@@ -105,7 +126,7 @@ h6 {
   color: rgb(252, 238, 248);
 }
 
-.registerGuestLink {
+.registerGuestLiank {
   cursor: pointer;
   color: rgb(218, 208, 214);
   font-size: 10px;
